python/database_api/requirements.txt
Cerberus==1.3.4
Flask-Cors==3.0.10
Flask-Migrate==3.1.0
Mako==1.2.2
SQLAlchemy==1.4.40
Six==1.16.0
alembic==1.8.1
certifi==2022.6.15
charset-normalizer==2.1.1
flask-marshmallow==0.14.0
flask-sqlalchemy==2.5.1
greenlet==1.1.3
gunicorn==20.1.0
idna==3.3
importlib-resources==5.9.0
marshmallow-sqlalchemy==0.28.1
marshmallow==3.17.1
pymysql==1.0.2
requests==2.28.1
responses==0.21.0
urllib3==1.26.12